app/requirements.txt
flask>=1
python-dotenv>=0.15
psycopg2-binary>=2.8
